+++ models/character_ownership.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Date, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Date, ForeignKey
 from database import Base
 import datetime
 
@@ -8,6 +8,6 @@ class CharacterOwnership(Base):
     
     id = Column(Integer, primary_key=True)
     character_id = Column(Integer, nullable=False)  # Livello.id
-    user_id = Column(Integer, nullable=False)  # Utente.id_telegram
+    user_id = Column(BigInteger, nullable=False)  # Utente.id_telegram
     equipped_at = Column(DateTime, nullable=False, default=datetime.datetime.now)
     last_change_date = Column(Date, nullable=True)  # Track for weekly restriction
