+++ models/dungeon.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, ForeignKey
 from database import Base
 import datetime
 
@@ -6,7 +6,7 @@ class Dungeon(Base):
     __tablename__ = "dungeon"
     id = Column(Integer, primary_key=True)
     name = Column(String, nullable=False)
-    chat_id = Column(Integer, nullable=False)
+    chat_id = Column(BigInteger, nullable=False)  # Telegram chat ID (BigInteger for large IDs)
     current_stage = Column(Integer, default=0)
     total_stages = Column(Integer, default=5)
     status = Column(String, default="registration") # registration, active, completed, failed
@@ -23,4 +23,4 @@ class DungeonParticipant(Base):
     __tablename__ = "dungeon_participant"
     id = Column(Integer, primary_key=True)
     dungeon_id = Column(Integer, ForeignKey('dungeon.id'))
-    user_id = Column(Integer, nullable=False) # Telegram ID
+    user_id = Column(BigInteger, nullable=False) # Telegram ID
