+++ models/stats.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, Float, DateTime, ForeignKey
 from database import Base
 import datetime
 
@@ -9,7 +9,7 @@ class UserStat(Base):
     """
     __tablename__ = "user_stat"
     
-    user_id = Column(Integer, primary_key=True)  # id_telegram
+    user_id = Column(BigInteger, primary_key=True)  # id_telegram
     stat_key = Column(String(50), primary_key=True)   # e.g., 'total_damage', 'total_kills'
     value = Column(Float, default=0.0)
     last_updated = Column(DateTime, default=datetime.datetime.now, onupdate=datetime.datetime.now)
